msm: pil-riva: Hold wakelock while proxy voting
If the system goes to suspend immediately after RIVA is brought
out of reset the proxy vote for CXO and the PLLs will stay
asserted until the kernel wakes up again. Since there is no
guarantee for the kernel to wakeup again from suspend for a very
long time we may block XO shutdown as well as keep regulators
enabled when they shouldn't be enabled.

To avoid such a situation register for a wakelock so that this
driver can block suspend until the allotted timeout has expired.
Doing this will ensure that the proxy vote is removed with a
minor hit to power while we wait for the timeout to expire.

Change-Id: Iff57c8834d05bb8d955c83b113c36877a72373d4
CRs-Fixed: 337631
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>