msm_xo: Work around RPM treating CXO_CLK and D0 as the same
On the RPM turning on the CXO_CLK resource turns on the D0
resource. Voting on the D0 resource doesn't do much besides
toggle the buffer on the PMIC. Since the D0_EN signal coming from
the MSM dictates if the buffer is on or not using the D0 buffer
resource is essentially a nop.

Remap the D0 buffer users to the CXO clock in the clock driver so
that the callers of the API don't have to deal with this problem.
Once the RPM properly separates the D0 buffer from the CXO_CLK
resource we can remove this.

Change-Id: I3d69fab415913489b70555e7e7e871000835f346
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>