msm: board-8930: Remove always_on=1 for PMIC PM8038 SMPS 1 and LDO 24
Change the configurations of PMIC PM8038 regulators SMPS 1 and
LDO 24 so that they have always_on=0 instead of always_on=1.
These regulators correspond to VDD_Dig and VDD_Mem respectively.

The always_on=1 constraint is no longer needed for these
regulators because they have been switched to RPM control and the
RPM ensures that voltage requirements of other processors are
also met.  This change is also needed to avoid an undesired S1
request for 1.15 V as VDD_Dig control is migrated to the new
voltage corner regulator.

Change-Id: I4a9f034447831991d15672014f6b558cdaf13d1b
Signed-off-by: David Collins <collinsd@codeaurora.org>