msm: qdss: rename ptm to etm
HW implementations for the ARM ETM specifications are referred to as ETM
(Embedded Trace Macrocell). Similarly HW implementations of the PFT
(Program Flow Trace) specifications are referred to as PTM (Program
Trace Macrocell).

Despite the above, ETM is a more popular name and commonly gets used to
refer to both ETM and PFT spec implementations. Moreover, PFT spec
register names use ETM as a qualifier.

Since it is possible, we will share the same driver for both PFT and ETM
spec implementations, just use the more familiar ETM as the name.

This change renames variables, strings, etc from PTM to ETM. This
completes the PTM to ETM renaming.

Change-Id: I95b6f041ef1988975ade534b5b503e23525db3a3
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>