power: pm8921-bms: approach fcc if outside the scope
The existing algorithm discards the learnt FCC (full charger capacity)
if it is differs by more than 3% of the original. It could happen that
a battery in field doesn't confirm to the FCC degradation of the
profiled battery. So instead of discarding, limit the new FCC to 3% of
the original if it is beyond that value.

Also update the algorithm to learn FCC values only if the charging
began when the battery was lower than 30%. This avoids learning
FCC when the battery was nearly full - avoiding a faulty value being
learnt.

Change-Id: I0bd939517eddcc2de702397628de3309d0855644
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>