msm: dma: Fix the error checking for DMA clocks.
If sometime while enabling dma clocks, clock operation fails, since we
are not sanity checking the error value and assumes that dma clocks are on.
Next time when dma timer expires, it tries to disable the dma clocks which
are not enabled earlier and we get a clock unbalance stack trace warnings.

This scenario is reproducible if we continuously reboot the target. If
clk enable is failed and we bail out immediately, the next timer event
will not be set and there will be no further call to clock disable.

CRs-Fixed: 342519
Change-Id: I68c07f35659ced09487035c246544da7abdf0c28
Signed-off-by: Pankaj Kumar <pakuma@codeaurora.org>