ASoC: wcd9304: Add run-time PM to codec startup and shutdown functions
The suspend currently goes on a blocking state due to the locks.
These locks have been moved to parent device in mfd core driver
for safe codec read write and hence no need for duplicate locking.

Add the run-time PM call to wcd9304 codec so Slimbus driver
keeps Slimbus clock on when the codec driver knows there is going
to be audio streams running on the codec.

Change-Id: Id3bf6ce9da5f7bd882c8c2bf927527d16f197bc0
Signed-off-by: Asish Bhattacharya <asishb@codeaurora.org>