msm: clock-9615: Remove dummy clock entries for ebi1
The adm bus voting clock was stubbed out until the RPM bus driver
was in place, otherwise the only request for bus bandwidth would
be done by the adm driver and ebi1 would drop to 0 when adm
wasn't in use. Now that the RPM driver is in place we can safely
remove these entries.

Change-Id: Ie5dd4593b7ccc22bfa8ed453c2f2eb629135e1b6
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>