mmc: msm_sdcc: Add mmc-async-req support
There are two optional members in the mmc_host_ops
  - pre_req
  - post_req
These two are implemented, in order to move work to
before and after the host's request method is called.
In the DMA case, the pre_req maps the sg_list of the
new request while the previous request is ongoing &
post_req unmaps the sg_list of the previous request
while the new request is ongoing.

Change-Id: Ic6298fa1380bb2b35391fbf3779e37fda824cf98
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>