msm: socinfo: Prepend target name to dummy socinfo
The build_id field of socinfo holds the target name information
that is used by some tests to detect the target and run
appropriate tests. In the absence of a boot loader, dummy values
are populated in the socinfo and absence of target name in the
build_id field results in incorrect test results.

Prepend copper target name in build_id field of dummy socinfo
to allow for the tests to detect the target in the absence of
boot loader.

Change-Id: I69d04012340e61c0f7428ec04487a01b5c60bc87
Signed-off-by: Sathish Ambley <sambley@codeaurora.org>