msm: audio: qdsp5v2: Fix LPA no audio heard issue
-In LPA Playback, insert headset just before end
of the clip, then no audio is heard and music
is not stopping.
-When the DSP side buffer consumption crosses the
threshold value of current buffer and requests
the next buffer and receives it from driver. At
this point of time, if device switch happens,
then driver gets sync info and flushes the DSP
data and DSP requests for needs data. Driver pops
the latest buffer that is not yet played by DSP
and remove from out_queue. Also, increments the
headcount to next buffer. This will miss the audio
play of few secs and result in failure of music stop
too as WRITE_DONE event to LPAPlayer is not sent
for one buffer.
-Apply the change to handle device switch case
to avoid removing the current buffer from out_queue
incase of device switch. Play the buffer and
then post the WRITE_DONE event of this buffer
to LPAPlayer. With this change music stop
happens correctly now as WRITE_DONE is issued
for all the buffers to LPAPlayer.

Change-Id: I7eb7ea4215da02e77a88a1f30bef53382f4f4faa
CRs-Fixed: 342560
Signed-off-by: Laxminath Kasam <lkasam@codeaurora.org>