usb: dwc3: add USB_DWC3_MSM peripheral controller choice.
We distinguish between the DWC3 generic modules, compiled
depends on USB_DWC3, to the vendor specific glue layers
as dwc3-msm.c or dwc3-omap.c complied by DWC3_MSM or DWC3_OMAP
configs. These configs automatically selects the USB_DWC3.

The USB_DWC3_MSM will be used in next commits for specific
needs of the dwc3 on MSM architecture.

Also removed USB_GADGET_DWC3 from Kconfig in order to organize
as in latest Linux kernel. This is done since the dwc3/gadget
is compiled anyway (as long as with the future dwc3/host) and only
on runtime (reading HWPARAMS) the dwc3/core decides if to load the
gadget (or the host).

Change-Id: Ifc2ebb1c9a6cfd59cd004aee5b528ba2bb364627
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>