mmc: msm_sdcc: use BAM for smaller SDCC transfers as well
BAM (DMA) can transfer the data of any length or alignment.
Currenly we are using the BAM only if SDCC transfer size
is less than FIFO_SIZE (64 bytes) and only if it's in multiple
of FIFO_SIZE.

As BAM doesn't have any limitation on the minimunm size of the
request, we can use the BAM even for transfers of less than
FIFO_SIZE but performance profiling has proved that using PIO
mode (no DMA) for transfers <= FIFO_SIZE is better compared to
BAM mode.

So this change uses the BAM only if transfer size is greater than
FIFO_SIZE but removes the restriction that transfer size must be
in multiple of FIFO_SIZE.

Change-Id: I41e4e43303200aac3860e33c96aeb78caf6031b5
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>