msm: smd: Add SMSM Reset Handshaking disable config
Add Subsystem Restart configuration to platform data that specifies if
the legacy SMSM_RESET Modem<->Apps handshaking should be used.  This
allows removing the calls to cpu_is_msmXXXX() which is currently being
used to disable the legacy handshaking.

Not disabling the handshaking for targets that support subsystem restart
will result in an erroneous modem-restart notification during a modem
subsystem restart which will typically result in a panic in the
subsystem restart handler.

Change-Id: Iaaf162f35c16d0abfff768ac62f00d3179dc5c91
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>