msm: dsps: Fix include path for timer.h
The C #include directive, when used with <> around the filename is not
supposed to look in the same directory as the source file containing
the directive, but only in the include path.  Some versions of the
compiler tools seem to not be obeying this, and msm_dsps.c relies on
this behavior, causing it to fail to compile with some compiler
versions.

Fix this by using "timer.h" instead of <timer.h>.

Change-Id: I4bb7316e335b832a936225ad7c22941b64bafd91
Signed-off-by: David Brown <davidb@codeaurora.org>