msm: bam_dmux: fix downlink data race condition
As soon as downlink data buffers are queued to the bam hardware, they can
be used.  Not all of the steps necessary to enable complete processing
of downlink data are completed when buffers are first queued.  This
creates a race condition where bam_dmux could attempt to process data
before it is ready.  Move the buffer queuing process to later when data
processing requirements are met, thus eliminating the race condition.

CRs-Fixed: 338616
Change-Id: I61b1bac5b66cd178141080d04907e742c9d9f81a
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>