Bluetooth: Wait for smd open before registering with hci
During subsystem restart we need to wait for the WCNSS to come up
and open the smd port before sending any data (command) to it
otherwise the data is lost and it will result in hci_cmd_timeout.

If there was a mechanism in place to issue subsystem_restart on
hci_cmd_timeout, it would result in repetitive subsystem_restart
without this fix.

CRs-fixed: 346227
Change-Id: I5b6098ad017455346953c63e6acef7b12eb44814
Signed-off-by: Sunny Kapdi <sunnyk@codeaurora.org>