qseecom:  Load and unload app on open/close
This fix is intended when using QSEOS_VERSION_13.
Targets that have not moved to latest QSEE, will run into
a failure when the qseecom daemon is run.   qseecom daemon
opens the qseecom device and registers the listeners.

In the old QSEE version 13, the app itself responds to these
request for registering the listeners.  So the app needs to have
loaded prior to issuing the register listeners request.

Change-Id: Ic1756aac2b8d8cc22b9b630053e77ba7f7a49c91
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>