diag: Fix incorrect parsing of log codes
DIAG maintains mask info in a buffer. This buffer is parsed
using a structure. Recently this structure was modified to add
another member. Now, the parsing code has to change to account
for this new member, otherwise, the parsing is incorrect and
misses certain log codes.

Also updated log codes with the latest on modem.

Change-Id: Iaaf86e4f3342fd73898eb11124b4893ec79aa564
Signed-off-by: Shalabh Jain <shalabhj@codeaurora.org>