msm: bam_dmux: send powerdown ack to A2
When bam_dmux disconnects to the bam hardware in response to the A2
powering down, send an ack back to the A2 signaling that bam_dmux got
the powerdown signal and processed it.

This change optimizes the response time for back-to-back shutdown and
powerup sequences instead of requiring a hard coded delay in the A2 state
machine.

Change-Id: Ie1699abb060f0c0ecd144060ec62955da9439288
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>