msm: Enable msm_adc for 7x27a/8x25
The msm_adc interface is used to read the HKADC from the
remote processor. It uses the DALRPC layer to communicate
to the remote HKADC driver.

Add support to read PMIC Thermistor from the HKADC to support
thermal clients on the apps. PMIC temperature is one of the
thermal nodes used for performing thermal mitigation.

Change-Id: I9d9e8178e19daf5d830588281855dd8a40de1f0d
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>