msm: pm-boot: Modify PM warmboot sequence instructions
Due to a h/w bug on 8625 once cores come out of the low
power mode(GDFS) and immediately executing the first
"ldr" instruction causes the fetching of improper value
from the memory. This causes the aborts on cores. This
patch adds few memory read instrucitons before it actually
jump to the warmboot entry function.

Change-Id: I76d1228878f3a5ea1179d51d043e0ebf1b92d2bc
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>