msm: modem-8960: Don't set SMSM_RESET when the modem crashes
When the modem crashes, it either sends an SMSM interrupt
with the modem SMSM_RESET bit set, or it triggers the
modem watchdog interrupt to the apps processor. Do not
set the apps SMSM_RESET bit in either case, as this will
trigger other subsystem drivers to restart their
individual subsystems.

CRs-fixed: 345622
Change-Id: I03680eedaa8f522c7727acd1e3fe92564c818d0d
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>