msm: rtb: Add rtb timestamps using sched_clock()
Provide timestamps in the rtb log for the purposes of correlating rtb
entries with dmesg logs. This provdes a 64 bit timestamp in nanoseconds,
overloading the caller field for the lower bits. In order to indicate to
parsing tools that the caller field is not a code pointer, the new entry
type has its highest order bit set.

Change-Id: I275b2a5981e6d0d61a658930914c37b63cfa3e5b
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>