msm: pm: clean boot vector in outer cache after writing
The boot vector will need to be cleaned in the outer cache
after being written to ensure that it can be read by a CPU
which is waking up from power collapse. Calling
v7_flush_dcache_all is not sufficient on targets which
have an outer cache.

There is currently no requirement for the boot vector write to
happen in an assembly environment, and moving it to C makes
cleaning the outer cache easier.

Change-Id: I64c2577c9159adf25b26995a7e3ac0285d01d0d6
Signed-off-by: Steve Muckle <smuckle@codeaurora.org>