msm: pm2: Program the top csr to put core1 into GDFS.
Upon wakeup source detection, modem bring's the core0 out of
reset. The recomendation for core0 in 8x25 is to put the core1
into GDFS till hotplug sw brings it online.

Since core1 is not yet enabled, program the top CSR from core0
context and drive the SPM1 to put the core1 in to GDFS. Once
core1 is put into GDFS disable the top csr and enable the SPM1
control for core1.

Enable the scu as well while coming out of power collapse which
facilitates the coherency between shared data across the cores.

change-Id: Iee35908d55c9a630da6bf2509a66a3c7f870fece
Signed-off-by: Anji jonnala <anjir@codeaurora.org>