hwmon: epm: Add file interface support
Add IOCTL to initialize the I2C expander and SPI
initialization for ADS ADC. Clients issue a request
to initialize and enable the device before issuing
requests to perform an ADC read.

The file interface is necessary since it takes
considerable amount of time to initialize the
I2C expander and the ADS ADC device for every
ADC read transaction.

Change-Id: I9e830232592288cbc41554070c70882f05c842ce
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>