msm: audio: qdsp6v2: PCM Encoder is configured to native channel count
Impact:
When the content switches from 5.1 channels to mono. Noise is heard
in one of the channels.

Problem:
At the start of the playback, number of output channels of PCM
Encoder is configured to 2 when the number of channels in the input
stream are greater than 2. This is done to confine maximum output
channels from PCM encoder to 2.

Change:
Modify output channel configuration of PCM encoder to native
channel count as Q6 decoder takes care for converting channel count
to 2 for the clips with channel count greater than 2.

Change-Id: I305e7c6d0e19977ae56503a8cdb0495231d51468
CRs-Fixed: 344426
Signed-off-by: Preetam Singh Ranawat <apranawat@codeaurora.org>