ASoC: wcd9310: Update the sequence to apply the digital gain
The digital gain implementation requires that the specific decimator or
interpolation path be enabled for the digital gain to be physically
applied in the signal path.  If the gain is set prior to enabling
the decimator/interpolator path, the register will show the programmed
gain value correctly, but it will not be applied in the signal
chain. Setting a digital gain value and then enabling the
decimator/interpolator path will not cause the gain to be applied.
The only way for the programmed gain value to be implemented
correctly is to first enable the decimator/interpolator path and then
set the digital gain of the path.

Change the codec driver to apply the digital gain again for a given
digital RX or TX path right after digital path is turned on.

CRS-fixed:  340236
Change-Id: Ie737a3875ad3a4e55a76ea17f155d02f6be4a85e
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>