Revert "USB: gadget: Enable streaming mode support in hardware"
This reverts commit f3e34e5d2b85478346bf3a826d2e0a74e67b97aa.

Reverting this change due to the following reason

Observing low throughputs on RNDIS DL with streaming mode enabled due to
inverted CRC errors on IN tokens on Windows 7 Host. This is resulting in
more time to transfer single IP packet, thereby accumulating to
degradation on overall throughputs of RNDIS interface.

Conflicts:

	drivers/usb/gadget/ci13xxx_msm.c

CRs-fixed: 332219
Change-Id: I36ce36a1ed78ee54ab00fc9c518a494a55f82fde
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>