msm: bam_dmux: pin rx workqueue to core 0
Pin the receive workqueue so that it always runs on core 0.  Do this to
eliminate a race condition caused by the workqueue moving between cores in
the middle of a downlink data transfer.  If hit, the race condition will
cause the TCP/IP framework to process packets out of order for a short time
which could cause performance degradation due to unnecessary
retransmissions.

CRs-Fixed: 346788
Change-Id: If1588cc5fd24f6d4af441a1728cfbb801a7d7c5b
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>