ASoC: wcd9310: Add support for MIC BIAS other than 1.8 volts
When MBHC is set to use microphone bias other than 1.8 volts, codec driver
needs to calculate two sets of thresholds. This is required because the
mic bias can be powered from two different sources. If vddio is used as
supply, then the maximum microphone bias voltage can only be 1.8 volts.
If MBHC originally was using other microphone voltage, the thresholds will
have to be changed when mic bias power is switched to vddio supply.

CRs-fixed: 341458
Change-Id: I428a4adaaa78b273cd598772c38bccdce8d9c27c
Signed-off-by: Joonwoo Park <joonwoop@codeaurora.org>