msm: board-copper: Add the ability to probe devices in order
Add a mechanism to allow for select devices to be probed early
and in order. This works by invoking specific driver init
routines in arch_init, before the vast majority of the drivers
are added in device_init. Invoke this new init routine after all
devices are added, and thus probe devices mapped to these
drivers before any others.

This approach does not solve the problem where multiple devices
using the same driver have dependencies between them.

Also add the stub-regulator driver to be the first member of this
new init routine. This driver supports devices that are required
early on.

Change-Id: Ib39e3111c8feb407eb8283b991ffe796dbbb5b90
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>