msm: smsm: Add support for SMSM interrupt mask
When a remote processor changes its SMSM state, it compares the change
against the notification mask and sends a notification to the remote
processor if needed.  Traditionally, Apps has set all interrupt flags to
enabled, so it is woken up and interrupted for all state changes.

With the recent addition of the SMSM callback notifiers, we can start
setting the mask based upon the callbacks.

CRs-Fixed: 347968
Change-Id: I84c06a00db5857d847ae318290df13babf317869
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>