usb: mdm_bridge: Fix premature auto suspend on Rx data path
mdm bridge driver does not prevent auto suspend while submitting
the URB on rx data path. This causes the usb device to go to low power
mode in the middle of the downlink data transfer as the auto suspend
timer does not get reset by the driver while submitting the URB and
it expires.

CRs-Fixed: 346331
Change-Id: Idd9c7066da7c9b4081d02126e3d0fc9e069adef5
Signed-off-by: Hemant Kumar <hemantk@codeaurora.org>