msm: Consolidate set-up of 'children' lists in msm_clock_init()
Calling clk_set_parent(clk, clk_get_parent(clk)) worked but is confusing
to read since it's not obvious why it's necessary to re-set a parent to
something that already appears to be the parent.

Make this clear by moving the similar implementations for associating
children with their parent from clock-voter.c and clock-local.c into
msm_clock_init(), since there is no reason this code can't be
generic.

Change-Id: Ibb064fd1f1dfd0696cfa8e3e3056ebc461c90540
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>