msm: dma: Convert spinlocks to mutexes for new clock apis
Converting the dma driver to mutexes is a requirement for using the new
clk_prepare and clk_unprepare apis. In order to deal with callers who
run in atomic context, refactor the enqueuing code to run in a
workqueue. Introduce a new spinlock that only protects the command lists
and not the hardware so the enqueue routine can add to the lists in
atomic context.

Change-Id: If681dece4421a0e28afa27a1aa7ccebf19f60056
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>