msm: 9615: Add support for ION
ION is a memory manager used by kernel and user
space clients to efficiently share buffers.

Add support for ION to 9615 target, as it is needed
for audio and power collapse. This commit will reserve
a total of 1mb of system memory (as reservations are
rounded up to megabyte boundaries).

A corresponding change to power management boot code
is also needed with this. When ION is enabled, the
method previously used by the PM code to allocate and
map the boot area for 9x15 does not work anymore. That
area must be allocated with a different API.

Change-Id: Id4354362dc9eccabd78943317a9e45d6844b3bd2
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>