board: 8960: Increase QSECOM ION heap to 7.5MB
When IOMMU is enabled the SF heap is not needed anymore. QSECOM was
using this heap in addition to the QSECOM heap and QSECOM
cannot use the IOMMU heap and thus we need to increase the
QSECOM heap size.

Increase QSECOM ION heap to 7.5MB from 6MB only when IOMMU is
enabled.

Change-Id: I29b6128c60953b39557ada23e6a3b2ed526f2840
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>