msm: clock-pll: Use a timeout when enabling voteable PLLs
Instead of waiting infinitely for a PLL to enable,
use a timeout mechanism that returns an error code
if that voteable PLL doesn't activate its output
in time.

This will especially help if PLLs are incorrectly
configured and cannot turn on during bootup. Rather
than the device hanging forever, the kernel log
will contain a WARN stack and the caller will
know that the PLL didn't enable.

Change-Id: I4eab4b72da7c28866f033b50121c8265c9d432ad
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>