msm: pm2: Handle modem early exit properly in PC for 8625
When the APPS decides to enter into power collapse(idle/suspend)
it executes the wfi instruction and generate A2M6_INT interrupt
to modem to remove the power rail. But there are cases where modem
receives an interrupt for the apps and uses M2A6_INT to wake up the
apps while the apps is trying to do a power collapse. But APPS has
executed the WFI instruction which causes SPM to put the core into
GDFS mode. Once APPS receives the M2A6_INT from modem it starts the
execution from the warm boot specficied by the core and return the
value as "1"(which tells PC is successful) even on modem early exit.

Currently this scenario is not properly handled for 8625 target and
these changes will differentiate between the modem early exit & full
power collapse.

Change-Id: I3b2bd37184b894a89197bc2097d671d2177ad437
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>